requirements.txt
scikit-learn
trimesh
tqdm
scipy
numpy
sqlalchemy
sshtunnel
pyopenssl
Flask
pandas
psycopg2
requests
wget
beautifulsoup4
anytree
gmsh-sdk
shapely